+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [1.0.2] - 2019-12-20
+### Added
+- Added promise function that returns instance of Promise for jueue. [README.md](https://github.com/JosephUz/jueue#jueuepromiselist-or-jueuepromise)
+- Unit test "use promise" is added. [Test File](https://github.com/JosephUz/jueue/blob/master/test/index.test.js#L342)
+- Unit test "catch error by using promise" is added. [Test File](https://github.com/JosephUz/jueue/blob/master/test/index.test.js#L360)
+
+### Changed
+- Readme updated for "promise". [README.md](https://github.com/JosephUz/jueue#jueuepromiselist-or-jueuepromise)
+
+
 ## [1.0.1] - 2019-12-20
 ### Added
 - Added "then" to catch called done in jueue. [README.md](https://github.com/JosephUz/jueue/blob/master/README.md#jueuethencb)